# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.CreateUserRequest:
    properties:
      address:
        type: string
      born_date:
        type: string
      email:
        type: string
      name:
        maxLength: 100
        minLength: 2
        type: string
      password:
        minLength: 6
        type: string
    required:
    - born_date
    - email
    - name
    - password
    type: object
...
  models.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  models.LoginResponse:
    properties:
      token:
        type: string
      user:
        $ref: '#/definitions/models.User'
    type: object
  models.UpdateUserRequest:
    properties:
      address:
        type: string
      born_date:
        type: string
      email:
        type: string
      name:
        maxLength: 100
        minLength: 2
        type: string
    type: object
  models.User:
    properties:
      address:
        type: string
      born_date:
        type: string
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      role:
        type: string
      updated_at:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: Tim Developer
    url: http://www.example.com/support
  description: Ini adalah dokumentasi API server untuk manajemen User, Buku, dan Peminjaman.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Gin Gorm API (Manajemen Buku)
  version: "2.0"
paths:
...
  /admin/stats:
    get:
      consumes:
      - application/json
      description: Menampilkan total user, user baru hari ini, dan 5 user terakhir.
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User Statistics (Total, New Today, Latest)
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden (Bukan Admin)
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Statistik User (Admin)
      tags:
      - admin
  /admin/users:
    get:
      consumes:
      - application/json
      description: Menampilkan daftar user dengan fitur pagination, limit, dan sorting.
        Khusus Admin.
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 'Halaman ke berapa (Default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Jumlah data per halaman (Default: 10, Max: 100)'
        in: query
        name: limit
        type: integer
      - description: 'Kolom untuk sorting (id, name, email, created_at, dll). Default:
          created_at'
        in: query
        name: sort_by
        type: string
      - description: 'Arah urutan (ASC atau DESC). Default: DESC'
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Response berisi data array user, objek pagination, dan sorting
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden (Bukan Admin)
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Lihat Semua User (Pagination & Sorting)
      tags:
      - admin
  /admin/users/search:
    get:
      consumes:
      - application/json
      description: Mencari data user berdasarkan keyword yang cocok dengan nama ATAU
        email (LIKE query).
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Keyword pencarian (Nama atau Email)
        in: query
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.User'
            type: array
        "400":
          description: Parameter query kosong
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Mencari User
      tags:
      - admin
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticates a user and returns a JWT token
      parameters:
      - description: Login Credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.LoginResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      summary: Login User
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Mendaftarkan user baru dengan nama, email, password, dan tanggal
        lahir.
      parameters:
      - description: Data Registrasi User
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Validasi Error / Format Tanggal Salah
          schema:
            additionalProperties: true
            type: object
        "409":
          description: Email sudah terdaftar
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Register User Baru
      tags:
      - auth
...
  /users/{id}:
    put:
      consumes:
      - application/json
      description: Mengubah data profil user (Nama, Alamat, Email, Tgl Lahir). Membutuhkan
        token login.
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: Data yang ingin diupdate
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: ID Salah / Format Tanggal Salah / Tidak ada field update
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: User tidak ditemukan
          schema:
            additionalProperties: true
            type: object
        "409":
          description: Email konflik dengan user lain
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update Data User
      tags:
      - users
  /users/profile:
    get:
      consumes:
      - application/json
      description: Menampilkan data detail user yang sedang login berdasarkan Token.
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "401":
          description: Token tidak ada / tidak valid
          schema:
            additionalProperties: true
            type: object
        "404":
          description: User tidak ditemukan
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Gagal mengambil data
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Profil Saya
      tags:
      - users
securityDefinitions:
  BearerAuth:
    description: Masukkan token dengan format "Bearer <token_anda>
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"